+++ api/src/main/java/net/jqwik/api/arbitraries/IteratorArbitrary.java
@@ -10,7 +10,7 @@ import static org.apiguardian.api.API.Status.*;
  * Fluent interface to add functionality to arbitraries that generate instances
  * of type {@linkplain Iterator}
  */
-@API(status = EXPERIMENTAL, since = "1.3.2")
+@API(status = MAINTAINED, since = "1.3.2")
 public interface IteratorArbitrary<T> extends StreamableArbitrary<T, Iterator<T>> {
 
 	/**