+++ api/src/main/java/net/jqwik/api/arbitraries/ListArbitrary.java
@@ -13,7 +13,7 @@ import static org.apiguardian.api.API.Status.*;
  * Fluent interface to add functionality to arbitraries that generate instances
  * of type {@linkplain List}
  */
-@API(status = EXPERIMENTAL, since = "1.3.2")
+@API(status = MAINTAINED, since = "1.3.2")
 public interface ListArbitrary<T> extends StreamableArbitrary<T, List<T>> {
 
 	/**
@@ -51,6 +51,7 @@ public interface ListArbitrary<T> extends StreamableArbitrary<T, List<T>> {
 	 * @param mapper Mapper function to element type U
 	 * @return arbitrary of a list of Us
 	 */
+	@API(status = EXPERIMENTAL, since = "1.2.1")
 	<U> Arbitrary<List<U>> mapEach(BiFunction<List<T>, T, U> mapper);
 
 	/**
@@ -61,5 +62,6 @@ public interface ListArbitrary<T> extends StreamableArbitrary<T, List<T>> {
 	 * @param flatMapper Mapper function to arbitrary of element type U
 	 * @return arbitrary of a list of Us
 	 */
+	@API(status = EXPERIMENTAL, since = "1.2.1")
 	<U> Arbitrary<List<U>> flatMapEach(BiFunction<List<T>, T, Arbitrary<U>> flatMapper);
 }