+++ docs/release-notes.md
@@ -53,6 +53,7 @@ and
     - `Arbitrary.set()` now has return type `net.jqwik.api.arbitraries.SetArbitrary`
     - `Arbitrary.list()` now has return type `net.jqwik.api.arbitraries.ListArbitrary`
     - `Arbitrary.stream()` now has return type `net.jqwik.api.arbitraries.StreamArbitrary`
+    - `Arbitrary.iterator()` now has return type `net.jqwik.api.arbitraries.IteratorArbitrary`
 
 - `net.jqwik.api.lifecycle.Reporter` moved to `net.jqwik.api.Reporter`